Bug fixes and bit of polish
P0 — Fixed:

mainwindow.cpp:416-419 — moduleChange() now removes module's frames (old), not mod's (new)
preferences.cpp:57-60 — removeListener() uses erase-remove idiom, no more iterator skip
bigfixedpoint.cpp:116-122 — separator-strip loop uses else ++i, matches the QString version

P1 — Fixed:

bigfixedpoint.cpp:108,135 — both string constructors now pass base=10 to mpz_class, preventing octal interpretation of leading zeros
bigfixedpoint.cpp:247 — pow() asserts power >= 0 and casts explicitly to unsigned long
bigfixedpoint.cpp:251-263 — root() asserts root > 0 and decimalPlaces % root == 0; sqrt() asserts decimalPlaces % 2 == 0; both cast to unsigned long
bigfixedpoint.cpp:506-555 — all /=, % operators add assert + runtime guard returning early on zero divisor or decimal-place contract violation

P3 — Fixed:

bigfixedpoint.h:154-157 — operator/ and operator% with int now take const BigFixedPoint&
mainwindow.h:46 / mainwindow.cpp:156 — getSeed() return type corrected to quint32

+++ mainwindow.h
@@ -43,7 +43,7 @@ public:
     ~MainWindow();
     void newQuestion();
     //void testSQL();
-    static quint64 getSeed();
+    static quint32 getSeed();
 
 protected:
     int answerGiven;
